requirements.txt
requests
beautifulsoup4
load_dotenv
